fix: extract shared preferences helpers, add readBody error handler, use readBody in mcp.ts
