+++ src/server/routes/utils.ts
@@ -1,5 +1,6 @@
 import http from "node:http";
 import type { Workflow } from "../agent/workflow-store.js";
+import type { PreferencesStore } from "../agent/preferences-store.js";
 
 /**
  * Parse the JSON body of an incoming request.
@@ -9,6 +10,7 @@ export function readBody(req: http.IncomingMessage): Promise<any> {
 	return new Promise((resolve) => {
 		const chunks: Buffer[] = [];
 		req.on("data", (chunk: Buffer) => chunks.push(chunk));
+		req.on("error", () => resolve(null));
 		req.on("end", () => {
 			try {
 				resolve(JSON.parse(Buffer.concat(chunks).toString()));
@@ -17,6 +19,26 @@ export function readBody(req: http.IncomingMessage): Promise<any> {
 			}
 		});
 	});
+}
+
+/** Return preferences with sensitive keys (providerKey.*) filtered out. */
+export function getSafePreferences(preferencesStore: PreferencesStore): Record<string, unknown> {
+	const all = preferencesStore.getAll();
+	const filtered: Record<string, unknown> = {};
+	for (const [key, value] of Object.entries(all)) {
+		if (!key.startsWith("providerKey.")) {
+			filtered[key] = value;
+		}
+	}
+	return filtered;
+}
+
+/** Broadcast preferences_changed with sensitive keys filtered out. */
+export function broadcastPreferencesChanged(
+	preferencesStore: PreferencesStore,
+	broadcastToAll: (msg: Record<string, unknown>) => void,
+): void {
+	broadcastToAll({ type: "preferences_changed", preferences: getSafePreferences(preferencesStore) });
 }
 
 /**
